requirements.txt
Django<4.0
black==20.8b1
djangorestframework<3.13
